apps/api/pyproject.toml
[tool.pdm]

[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "Antonio Reid", email = "read.musik@gmail.com"},
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "auth0-python>=3.2.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
